pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai-policy-lab"
version = "0.1.0"
description = "Multi-agent policy research lab for graduate-level AI labor market analysis."
authors = ["Mary <mfl2147@columbia.edu>"]
readme = "README.md"
packages = [{ include = "ai_policy_lab", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
httpx = ">=0.27.0,<1.0.0"
langgraph = ">=0.2.0,<1.0.0"
langchain-core = ">=0.3.0,<1.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
rich = ">=13.9.0,<15.0.0"
typer = ">=0.12.5,<1.0.0"

[tool.poetry.group.dev.dependencies]
mypy = ">=1.11.2,<2.0.0"
pytest = ">=8.3.3,<9.0.0"
pytest-cov = ">=5.0.0,<6.0.0"
ruff = ">=0.6.9,<1.0.0"

[tool.poetry.scripts]
ai-policy-lab = "ai_policy_lab.cli:app"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008", "E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
exclude = ["tests/"]
